lf - open editor with key press with bat preview
This feature itself is not implemented to bat just yet
but mentioned in this issue - https://github.com/sharkdp/bat/issues/1053.

Fortunately though I was able to use the proof-of-concept library
by the creator of the bat, https://github.com/sharkdp/riffle
and set it up to work for my need (press "e" to open editor)
and this commit is the source code of how to make it work
while waiting this feature to be ready on bat itself in the future.