nvim - revive tidy.nvim
For sometimes, tidy.nvim stopped working and today I found out why.
It's because I started using .editorcofig and that one didn't include
the option of `trim_trailing_whitespace`.

Now I fix this issue adding the options for both .editorconfig and tidy
(explained the why in the code as a comment)